DEVLOG 3 (Adds quivers item, no functionality)
DEVLOG 3!!

Added the quiver item, made the texture based on the original but touched it up a bit to match with the rest of the game since the texture update.

Working on behaviour. Plan is to make them behave like bundles for arrows, holding any assortment up to 128.

Java is a pain to learn